requirements.txt
Click==7.0
Flask==1.0.2
flask-login==0.4.1
flask-migrate==2.3.1
flask-sqlalchemy==2.3.2
flask_wtf==0.14.2
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.4.0
psycopg2==2.7.6
PyMySQL==0.9.3
PyYAML==3.13
redis==2.10.6
requests==2.21.0
rq==0.12.0
Werkzeug==0.14.1
